Redesign constructor signature
This redesign features an options object, which is more flexible and forward-compatible. It also makes the registry a two-way mapping, which allows automatic lookup of the appropriate (local name, namespace) pairs---at least when there is only one of them---inside the base constructor. That enables `new HTMLParagraphElement()` to work, without having to pass it a local name. The same logic applies to custom elements; see e.g. https://lists.w3.org/Archives/Public/public-webapps/2015JanMar/0230.html.

The net win of this design is that classes that derive from HTMLElement never have to override their constructor or `[Symbol.species]`.